backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.0=py38haa95532_0
aiohttp=3.10.5=py38h827c3e9_0
aiosignal=1.2.0=pyhd3eb1b0_0
annotated-types=0.6.0=py38haa95532_0
anyio=4.2.0=py38haa95532_0
appdirs=1.4.4=pyhd3eb1b0_0
async-timeout=4.0.3=py38haa95532_0
attrs=23.1.0=py38haa95532_0
blas=1.0=mkl
bottleneck=1.3.7=py38h9128911_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
brotli-python=1.0.9=py38hd77b12b_8
ca-certificates=2024.9.24=haa95532_0
certifi=2024.8.30=py38haa95532_0
cffi=1.17.1=py38h827c3e9_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py38haa95532_0
colorama=0.4.6=py38haa95532_0
contourpy=1.0.5=py38h59b6b97_0
cryptography=43.0.0=py38h89fc84f_0
cycler=0.11.0=pyhd3eb1b0_0
distro=1.9.0=pypi_0
dnspython=2.4.2=py38haa95532_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
email-validator=2.2.0=pyhd8ed1ab_0
email_validator=2.2.0=hd8ed1ab_0
et_xmlfile=1.1.0=py38haa95532_0
exceptiongroup=1.2.0=py38haa95532_0
fastapi=0.115.0=pyhd8ed1ab_0
fastapi-cli=0.0.5=pyhd8ed1ab_0
fonttools=4.51.0=py38h2bbff1b_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=py38h2bbff1b_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.43=py38haa95532_0
h11=0.14.0=py38haa95532_0
httpcore=1.0.2=py38haa95532_0
httpx=0.27.0=py38haa95532_0
icc_rt=2022.1.0=h6049295_2
idna=3.7=py38haa95532_0
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.4=py38haa95532_0
jiter=0.6.1=pypi_0
joblib=1.4.2=py38haa95532_0
jpeg=9e=h827c3e9_3
kiwisolver=1.4.4=py38hd77b12b_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libabseil=20240116.2=cxx17_h5da7b33_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=4.25.3=hf2fb9eb_0
libtiff=4.5.1=hd77b12b_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_1
markdown-it-py=2.2.0=py38haa95532_1
markupsafe=2.1.3=py38h2bbff1b_0
matplotlib-base=3.6.2=py38h1094b8e_0
mdurl=0.1.0=py38haa95532_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
multidict=6.0.4=py38h2bbff1b_0
numexpr=2.8.4=py38h5b0cc5e_0
numpy=1.24.3=py38hf95b240_0
numpy-base=1.24.3=py38h005ec55_0
openai=1.51.2=pypi_0
openjpeg=2.5.2=hae555c5_0
openpyxl=3.1.5=py38h827c3e9_0
openssl=3.0.15=h827c3e9_0
packaging=24.1=py38haa95532_0
pandas=2.0.3=py38h4ed8f06_0
pandas-stubs=1.2.0.62=py38haa244fe_0
pillow=10.4.0=py38h827c3e9_0
pip=24.2=py38haa95532_0
platformdirs=3.10.0=py38haa95532_0
plotly=5.24.1=py38h9909e9c_0
pooch=1.7.0=py38haa95532_0
protobuf=4.25.3=py38h958608f_0
psutil=5.9.0=py38h2bbff1b_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.8.2=py38haa95532_0
pydantic-core=2.20.1=py38hefb1915_0
pygments=2.15.1=py38haa95532_1
pyparsing=3.1.2=py38haa95532_0
pysocks=1.7.1=py38haa95532_0
python=3.8.20=h8205438_0
python-dateutil=2.9.0post0=py38haa95532_2
python-dotenv=1.0.1=pyhd8ed1ab_0
python-multipart=0.0.9=py38haa95532_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytz=2024.1=py38haa95532_0
pyyaml=6.0.1=py38h2bbff1b_0
requests=2.32.3=py38haa95532_0
rich=13.7.1=py38haa95532_0
scikit-learn=1.3.0=py38h4ed8f06_1
scipy=1.10.1=py38h321e85e_0
sentry-sdk=1.9.0=py38haa95532_0
setproctitle=1.2.2=py38h2bbff1b_1004
setuptools=75.1.0=py38haa95532_0
shellingham=1.5.0=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
sniffio=1.3.0=py38haa95532_0
sqlite=3.45.3=h2bbff1b_0
starlette=0.38.2=py38haa95532_0
tenacity=8.2.3=py38haa95532_0
threadpoolctl=3.5.0=py38h9909e9c_0
tqdm=4.66.5=py38h9909e9c_0
typer=0.12.5=pyhd8ed1ab_0
typer-slim=0.12.5=pyhd8ed1ab_0
typer-slim-standard=0.12.5=hd8ed1ab_0
typing-extensions=4.11.0=py38haa95532_0
typing_extensions=4.11.0=py38haa95532_0
unicodedata2=15.1.0=py38h2bbff1b_0
urllib3=2.2.3=py38haa95532_0
uvicorn=0.20.0=py38haa95532_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wandb=0.16.6=pyhd8ed1ab_1
wheel=0.44.0=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.11.0=py38h827c3e9_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
